+++ lib/handler/message_handler.dart
@@ -1,14 +1,13 @@
 import 'dart:convert';
 
-import 'package:connectars/config.dart';
-import 'package:connectars/connectars.dart';
-import 'package:connectars/dao/client.dart' as dao;
+import 'package:connectars/dao/client.dart' as client;
 import 'package:connectars/message/generic_message.dart';
+import 'package:connectars/service/config.dart';
 import 'package:connectars/service/log.dart';
 import 'package:connectars/service/pusher.dart';
 import 'package:http/http.dart';
 
-Future<void> handle(data, dao.Client client) async {
+Future<void> handle(data, client.Client client) async {
   var incomingMap = jsonDecode(data) as Map<String, dynamic>;
 
   LogService().log('[I] ' + incomingMap.toString());
@@ -17,14 +16,16 @@ Future<void> handle(data, dao.Client client) async {
 
   if (incoming.id == 'pong') {
     client.isAlive = true;
-    pingPongClient(client);
     return;
   }
 
-  var url = Config.API_BASE + Config.API_ROUTE_MESSAGE;
+  var url = ConfigService().config.API_BASE +
+      ConfigService().config.API_ROUTE_MESSAGE;
 
-  var response = await post(url,
-      body: {Config.API_ROUTE_MESSAGE_PARAMETER: jsonEncode(incoming.toMap())});
+  var response = await post(url, body: {
+    ConfigService().config.API_ROUTE_MESSAGE_PARAMETER:
+        jsonEncode(incoming.toMap())
+  });
 
   if (response.statusCode >= 200 && response.statusCode < 400) {
     LogService().log('[O] ' + response.body);