+++ lib/handler/request_handler.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:connectars/config.dart';
 import 'package:connectars/message/generic_message.dart';
+import 'package:connectars/service/config.dart';
 import 'package:connectars/service/log.dart';
 import 'package:connectars/dao/client.dart';
 import 'package:connectars/dao/connections.dart';
@@ -12,10 +12,11 @@ import 'package:http/http.dart' as http;
 Future<Client> connect(HttpRequest request) async {
   var token = request.headers.value('x-auth-token');
 
-  var url = Config.API_BASE + Config.API_ROUTE_AUTH;
+  var url =
+      ConfigService().config.API_BASE + ConfigService().config.API_ROUTE_AUTH;
 
-  var response =
-      await http.post(url, headers: {Config.API_ROUTE_AUTH_HEADER: token});
+  var response = await http.post(url,
+      headers: {ConfigService().config.API_ROUTE_AUTH_HEADER: token});
 
   String uuid;
   if (response.statusCode >= 200 && response.statusCode < 400) {