+++ lib/service/log.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:connectars/config.dart';
+import 'package:connectars/service/config.dart';
 
 class LogService {
   factory LogService() {
@@ -16,9 +16,10 @@ class LogService {
   static const typeRequest = 'request';
 
   void log(String log, {String type = typeMessage}) async {
-    await Directory(Config.SOCKET_LOG_PATH).create(recursive: true);
+    await Directory(ConfigService().config.SOCKET_LOG_PATH)
+        .create(recursive: true);
 
-    final filename = Config.SOCKET_LOG_PATH +
+    final filename = ConfigService().config.SOCKET_LOG_PATH +
         '/${DateTime.now().year.toString()}-${DateTime.now().month.toString()}-${DateTime.now().day.toString()}-$type';
 
     File(filename).writeAsStringSync('$log\n', mode: FileMode.append);